add GetCardsList component
