+++ src/mixins/filters.js
@@ -1,0 +1,11 @@
+
+export default {
+  filters: {
+    toUppercase(val) {
+      return val.toUpperCase()
+    },
+    snippet(val) {
+      return val.slice(0, 200) + '...'
+    }
+  },
+}
